Criei arquivos category_view.dart, promocoes_view.dart, e apaguei listCategory.md e listPromotion.md
